Style the list and hyperlinks in index.html
I added the recipe class and an id that correlates with the name of the recipe to each of the recipes included in the unordered list on the homepage. I styled each of the recipe names, which were items in the list, by increasing the font size; changing the font family; adjusting the spacing of each list item; adding a red background color, black border, an image of the recipe at the end of the list item's width on the right. I also styled the hyperlinks by changing the font color to black and removing the underline when the mouse is not hovering over a hyperlink. When the mouse is hovering over a hyperlink, the hyperlink's font color changes to white and a wavy underline appears under the text.