Adjust CSS styling for mobile devices
If a device's screen size is less than 765 pixels in width, a few CSS rules will be adjusted to improve the homepage's user friendliness on devices with smaller screens. These adjustments include moving the thickest border to the top instead of the left of the div element; ensuring that the width is set to "fit-content"; reducing the margins, width of the hyperlinks in the list's width, and size of the images of the meals; and removing the box-shadow.