Create a div around conch-fritters.html content and style
I placed a div element around the headings, image, paragraph, unordered list, and ordered list on the conch-fritters.html page. Then, I added a black border and white background color to the div, adjusted the spacing of the div on the homepage, and added a box shadow to the div.