Style the description, ingredients, and steps headings in conch-fritters.html
I added an id that correlates with the name of the heading to the ingredients and steps headings on the conch-fritters.html page. I styled each of these headings by increasing the font size, adjusting the spacing and font style of these headings, aligning the text of these headers, and adding an underline under these headers.