Style the description, ingredients, and steps sections in the card in conch-fritters.html
I added the color class to the description, ingredients, and steps sections in the card in conch-fritters.html. I styled each of these sections by changing the background color to a light orange color, changing the font style of the paragraph and the bullet point style of the unordered list, and adjusting the spacing of each section. I also reduced the font size of the description, ingredients, and steps headings so that the contents of the page would fit on one page.