Adjust CSS styling for mobile devices
If a device's screen size is less than 765 pixels in width, a few CSS rules will be adjusted to improve the conch-fritters.html page's user friendliness on devices with smaller screens. These adjustments include reducing the margins of the card, adding padding to the card, ensuring that the width of the card and the two columns are 100%, and removing the box-shadow.