Style the lobster-alfredo.html page
I applied the structure of the conch-fritters.html page to the lobster-alfredo.html page and the styling of the conch-fritters.css to the lobster-alfredo.css so that each recipe card can have a similar design. I adjusted the background color of the body and the description, ingredients, and steps sections in the lobster-alfredo.css.