Style the shrimp-and-sausage-gumbo.html page
I applied the structure of the lobster-alfredo.html page to the shrimp-and-sausage-gumbo.html page and the styling of the lobster-alfredo.css to the shrimp-and-sausage-gumbo.css so that each recipe card can have a similar design. I adjusted the background color of the body and the description, ingredients, and steps sections in the shrimp-and-sausage-gumbo.css.